Moved thread pool tests from shared to ruby-specific.
The specs for the Java-specific thread pools are mainly intended to
1) ensure consistent APIs and 2) verify that the Ruby thread pools
have matching behavior. Some of the more complex behaviors (such
as thread creation and garbage collection) are difficult to test
for the Java thread pools. In these cases we'll just trust that
Java's thread pools behave as documented and just test the Ruby
versions.