Add ReentrantReadWriteLock
ReadWriteLock is a fast and very useful utility, but it is non-reentrant and this
makes it unsuitable for some (well, a lot of) applications. You can never take an
exclusive lock without being 100% sure that you are not holding it already. If you
mess up, then game over. Your thread will deadlock hard and grind to a halt. Any
other thread which tries to take the exclusive lock will deadlock too.

When there are only a few sections of code which access a shared, mutable resource,
avoiding this scenario is not so hard. But when you have a lot of places where
locking is needed, and they call each other in complex ways, forget it. You WILL
get it wrong.

If you are building a library for other programmers to use, and you call into
"consumer" code with locks held, that's even worse. (Such as public API methods
which yield to a block...) You don't know what that "consumer" code will do. It
may call back into your library again.

Reentrant locks are so much more friendly. You can audit each method
individually to check whether it acquires (and releases) all the locks that
it needs in the right places. You don't have to check every single call site to
see which locks may already be held when the method is entered. (This will save you
some lost hair, which you can later tear out when dealing with lock acquisition
order issues.)

Acquiring and releasing a ReentrantReadWriteLock is about 50% more expensive
(in CPU time) than a plain ReadWriteLock. But it's still fast. Don't worry
about performance unless you expect to acquire/release it 10,000s of times
per second.

ReentrantReadWriteLock has another nice property: if you are holding a write
lock, you can FIRST acquire a read lock and THEN drop the write lock. This allows
lock "downgrading". "Upgrading" can be done in the converse way.

If you think "ReentrantReadWriteLock" is too long to type (and it is), feel free
to "alias" it by defining a constant. This is Ruby, after all.