Temporarily adding Rbx to allowed-failures list.
As a rule we support Rbx. We even have several Rbx optimizations. For
some unknown reason we are having intermittent test failures on Rbx that
make no sense. The stack tracs will show a file related to channels but
RSpec will show a test related to thread pools. It's nearly impossible
to determine what's happening when the stack trace doesn't match the
failing test.

This change is only temporary. As soon as I figure out what's going on
I'll remove Rbx from the allowed failures list.