Add CI flow for tests on Github actions
Travis CI has been down for a while now and we'd like to switch
to Github Actions. This change introduces a new CI build on Github
Actions that runs tests on Ruby implementations and also mark
some Ruby failing implementations as experimental so the build
still passes.

It does not contain all the requirements from the previous Travis
CI yet, but it's a good start.